update execution of pipeline code from a primitive
