+++ src/main/groovy/org/boozallen/plugins/jte/hooks/Hooks.groovy
@@ -24,6 +24,7 @@ import java.util.ArrayList
 import java.lang.annotation.Annotation
 import org.jenkinsci.plugins.workflow.cps.CpsThread 
 import org.jenkinsci.plugins.scriptsecurity.sandbox.whitelists.Whitelisted 
+import jenkins.model.Jenkins
 
 class Hooks implements Serializable{
 
@@ -48,20 +49,17 @@ class Hooks implements Serializable{
     @Whitelisted
     static void invoke(Class<? extends Annotation> a, TemplateBinding b, Map context = [:]){
         /*
-
+            any invocation of pipeline code from a plugin class of more than one
+            executable script or closure requires to be parsed through the execution
+            shell or the method returns prematurely after the first execution
         */
-        Utils.parseScript("""
-        import org.boozallen.plugins.jte.hooks.Hooks
-        import org.boozallen.plugins.jte.binding.TemplateBinding
-        import org.boozallen.plugins.jte.hooks.AnnotatedMethod
-        import java.util.ArrayList
-        import java.lang.annotation.Annotation
-         
-        void call(Class<? extends Annotation> a, TemplateBinding b, Map context){
-            List<AnnotatedMethod> discovered = Hooks.discover(a, b)            
-            discovered.each{ it.invoke(context) }
-        }
-        """, b)(a, b, context)
+        String invoke =  Jenkins.instance
+                                .pluginManager
+                                .uberClassLoader
+                                .loadClass("org.boozallen.plugins.jte.hooks.Hooks")
+                                .getResource("Invoke.groovy")
+                                .text
+        Utils.parseScript(invoke, b)(a, b, context)
     }
     
 }
