+++ src/main/resources/org/boozallen/plugins/jte/binding/Stage.groovy
@@ -1,0 +1,10 @@
+import org.jenkinsci.plugins.workflow.cps.CpsScript
+import org.codehaus.groovy.runtime.InvokerHelper
+import org.codehaus.groovy.runtime.InvokerInvocationException
+
+def call(CpsScript script, ArrayList<String> steps){
+    for(def i = 0; i < steps.size(); i++){
+        String step = steps.get(i)
+        InvokerHelper.getMetaClass(script).invokeMethod(script, step, null)
+    }
+}
